Update with gem structure.
