+++ README.md
@@ -1,0 +1,45 @@
+# Skeleton
+
+Create new apps quickly!
+
+## Installation
+
+Install this gem as follow:
+
+    $ gem install datacrafts-io-skeleton
+
+## Usage
+
+### New Rails API application without frontend part.
+
+This command will create new rails app with configured rails_helper.rb and .rubocop.yml files. All major gems are also included.
+
+    $ datacrafts-io-skeleton create APP_NAME
+    $ datacrafts-io-skeleton build APP_NAME
+    $ datacrafts-io-skeleton new APP_NAME
+    
+### With frontend part.
+
+You can specify `--frontend (-f) FRONTEND_TYPE` option to additionally create frontend application inside `APP_NAME/frontend` directory. Use `-t` option to add typescript support.
+
+    $ datacrafts-io-skeleton create APP_NAME --frontend react -t 
+
+Allowed frontend types include only `react` for now.
+
+## Development
+
+After checking out the repo, run `bin/setup` to install dependencies. Then, run `rake spec` to run the tests. You can also run `bin/console` for an interactive prompt that will allow you to experiment.
+
+To install this gem onto your local machine, run `bundle exec rake install`.
+
+## Contributing
+
+Bug reports and pull requests are welcome on GitHub at [datacrafts-io/skeleton](https://github.com/datacrafts-io/skeleton). This project is intended to be a safe, welcoming space for collaboration, and contributors are expected to adhere to the [code of conduct](https://github.com/datacrafts-io/skeleton/blob/master/CODE_OF_CONDUCT.md).
+
+## License
+
+The gem is available as open source under the terms of the [MIT License](https://opensource.org/licenses/MIT).
+
+## Code of Conduct
+
+Everyone interacting in the Skeleton project's codebases, issue trackers, chat rooms and mailing lists is expected to follow the [code of conduct](https://github.com/datacrafts-io/skeleton/blob/master/CODE_OF_CONDUCT.md).
