--- config.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-module Config
-  ROOT_PATH = File.expand_path(".")
-end
