--- helpers.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-module Helpers
-  def app(path=nil)
-    "#{app_name}/#{path}"
-  end
-
-  def commit(message)
-    run "git add . && git commit -m '#{message}'"
-  end
-end
