+++ lib/datacrafts_io_skeleton/config.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module DatacraftsIoSkeleton
+  module Config
+    ROOT_PATH = File.expand_path(File.join(__dir__, "..", ".."))
+  end
+end
