+++ lib/datacrafts_io_skeleton/version.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module DatacraftsIoSkeleton
+  # Defines gem version
+
+  VERSION = "1.0.0"
+end
