+++ spec/datacrafts_io_skeleton_spec.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+RSpec.describe DatacraftsIoSkeleton do
+  it "has a version number" do
+    expect(DatacraftsIoSkeleton::VERSION).not_to be_nil
+  end
+end
