+++ templates/rails.rb.erb
@@ -1,19 +1,19 @@
-<%= Gemfiler.render(:common) %>
+<%= DatacraftsIoSkeleton::Gemfiler.render(:common) %>
 
 gem_group :development do
-  <%= Gemfiler.render(:development) %>
+  <%= DatacraftsIoSkeleton::Gemfiler.render(:development) %>
 end
 
 gem_group :development, :test do
-  <%= Gemfiler.render(:development_test) %>
+  <%= DatacraftsIoSkeleton::Gemfiler.render(:development_test) %>
 end
 
 gem_group :test do
-  <%= Gemfiler.render(:test) %>
+  <%= DatacraftsIoSkeleton::Gemfiler.render(:test) %>
 end
 
-<% if Gemfiler.uncommented_default_gems.any? %>
-  uncomment_lines "Gemfile", <%= Gemfiler.uncommented_default_gems.map { |i| "\"#{i}\"" }.join(", ") %>
+<% if DatacraftsIoSkeleton::Gemfiler.uncommented_default_gems.any? %>
+  uncomment_lines "Gemfile", <%= DatacraftsIoSkeleton::Gemfiler.uncommented_default_gems.map { |i| "\"#{i}\"" }.join(", ") %>
 <% end %>
 
 run "bundle install"
